Refactor: Simplify favicon setting logic
I encapsulated favicon setting logic into a new `setFavicon` function in `src/UTILS/favicon.ts`.
Then, I updated `src/index.ts` to use this new function.

This change simplifies the favicon handling in the main entry point and makes the logic more reusable and maintainable.
I also removed browser-specific hacks in favor of a more standard approach.
